requirements.txt
flask==3.0.0
waitress==2.1.2
faiss-cpu==1.13.1
sentence-transformers==5.2.0
huggingface-hub==0.36.0
torch==2.9.1
openai==2.14.0
python-dotenv==1.0.0
pypdf==3.17.0
llama-cpp-python==0.3.2
ragas==0.4.2
scikit-learn==1.8.0
datasets==4.4.2
langchain==1.2.0
langchain-openai==1.1.6
pydantic==2.12.5
numpy==2.4.0
requests==2.32.5
